plugins/example/import/google-calendar/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
requests==2.31.0
python-dotenv==1.0.0